made constraint update/remove/add functions generic
