+++ src/main/webapp/app/modules/administration/user-management/permissions/permissions-util.ts
@@ -37,3 +37,29 @@ export const generateFeatureNameOptions = featurs => {
     });
   return options;
 };
+
+export const resetConstraint = (constraint, condition) => {
+  constraint.constraintDefinition.featureConstraints[0] = condition;
+  return Object.assign({}, constraint);
+};
+
+export const updateConditionInConstraint = (constraint, condition) => {
+  const index = constraint.constraintDefinition.featureConstraints.indexOf(condition);
+  if (index > -1) {
+    constraint.constraintDefinition.featureConstraints[index] = condition;
+  }
+  return Object.assign({}, constraint);
+};
+
+export const removeConstraintFromList = (constraint, condition) => {
+  const index = constraint.constraintDefinition.featureConstraints.indexOf(condition);
+  if (index > -1) {
+    constraint.constraintDefinition.featureConstraints.splice(index, 1);
+  }
+  return Object.assign({}, constraint);
+};
+
+export const addConstraintIntoList = (constraint, conditionDefaultValue) => {
+  constraint.constraintDefinition.featureConstraints.push(conditionDefaultValue);
+  return Object.assign({}, constraint);
+};
